[♻️refactor]: Icon export문 수정 및 Icon 컴포넌트 교체
- PostSimpleCard Icon 컴포넌트 교체
- 좋아요 버튼 누를 때 Favorite 글자가 뜨는 오류 수정
