+++ src/components/common/Navigator/TopNavBar/Elements/BackButton.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { styled } from "styled-components";
-import Icon from "../../../Icon/Icon";
+import { Icon } from "@/components/common";
 import { BACK_ICON } from "@/constants/icons";
 
 interface IBackButtonProps {
